+++ systemimage/tests/test_config.py
@@ -28,7 +28,7 @@ from pkg_resources import resource_filename
 from systemimage.config import Configuration, config
 from systemimage.reboot import Reboot
 from systemimage.scores import WeightedScorer
-from systemimage.tests.helpers import test_data_path, testable_configuration
+from systemimage.testing.helpers import test_data_path, testable_configuration
 from systemimage.tests.reboot import TestableReboot