+++ systemimage/tests/test_state.py
@@ -30,7 +30,7 @@ from systemimage.config import config
 from systemimage.gpg import SignatureError
 from systemimage.logging import initialize
 from systemimage.state import State
-from systemimage.tests.helpers import (
+from systemimage.testing.helpers import (
     copy, get_index, make_http_server, setup_index, setup_keyring_txz,
     setup_keyrings, sign, temporary_directory, testable_configuration)
 
@@ -428,9 +428,11 @@ class TestRebooting(unittest.TestCase):
         with unittest.mock.patch(
                 'systemimage.tests.reboot.TestableReboot.reboot', reboot_mock):
             list(State())
+        self.assertTrue(got_reboot)
 
     @testable_configuration
     def test_reboot_command_file(self):
+        # The command file gets properly filled.
         self._setup_keyrings()
         list(State())
         path = os.path.join(config.updater.cache_partition, 'ubuntu_command')
@@ -444,3 +446,35 @@ update 6.txt 6.txt.asc
 update 7.txt 7.txt.asc
 update 5.txt 5.txt.asc
 """)
+
+    @testable_configuration
+    def test_run_until(self):
+        # It is possible to run the state machine either until some specific
+        # state is completed, or it runs to the end.
+        self._setup_keyrings()
+        state = State()
+        self.assertIsNone(state.channels)
+        state.run_thru('get_channel')
+        self.assertIsNotNone(state.channels)
+        # But there is no index file yet.
+        self.assertIsNone(state.index)
+        # Run it some more.
+        state.run_thru('get_index')
+        self.assertIsNotNone(state.index)
+        # Run until just before the reboot.
+        #
+        # Mock the reboot to make sure a reboot did not get issued.
+        got_reboot = False
+        def reboot_mock(self):
+            nonlocal got_reboot
+            got_reboot = True
+        with unittest.mock.patch(
+                'systemimage.tests.reboot.TestableReboot.reboot', reboot_mock):
+            state.run_until('reboot')
+        # No reboot got issued.
+        self.assertFalse(got_reboot)
+        # Finish it off.
+        with unittest.mock.patch(
+                'systemimage.tests.reboot.TestableReboot.reboot', reboot_mock):
+            list(state)
+        self.assertTrue(got_reboot)